requirements.txt
aiohttp[speedups]==3.6.2
colored==1.4.2
marshmallow==3.5.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.0.1
paramiko==2.7.1
pycryptodome==3.9.7
pymongo==3.10.1
sentry_sdk[flask]==0.14.2
git+https://github.com/bwind/sticky-marshmallow.git
